transfermanager: fix Integer vs Long bug in CopyManager
Motivation:

Commit afd02360595 changed the storage type for active transfers from
Long to Integer.  Unfortunately, there was one place this patch failed
to update, which broke the 'ls internal' admin command and resulted in
a HIGH level SpotBugs warning.

Modification:

Parse admin command as an integer.

Result:

An unreleased regression is fixed with the 'ls internal' admin command.
A HIGH level SpotBugs warning is removed.

Target: master
Requires-notes: no
Requires-book: no
Patch: https://rb.dcache.org/r/12592/
Acked-by: Tigran Mkrtchyan

+++ modules/dcache/src/main/java/diskCacheV111/doors/CopyManager.java
@@ -91,7 +91,7 @@ public class CopyManager extends AbstractCellComponent
     {
         boolean long_format = args.hasOption("l");
         if (args.argc() > 0) {
-            long id = Long.parseLong(args.argv(0));
+            int id = Integer.parseInt(args.argv(0));
             CopyHandler handler = _activeTransfers.get(id);
             return id + ": " + (handler == null ? "no such ID" : handler.toString(long_format));
         }
